pyproject.toml
[project]
name = "Ameriabank"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.17",
    "dagster-docker>=0.28.17",
    "dagster-duckdb>=0.28.17",
    "dagster-k8s>=0.28.17",
    "dagster-polars>=0.27.9",
    "docker>=7.1.0",
    "kubernetes>=35.0.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "Ameriabank"
registry_modules = [
    "Ameriabank.components.*",
]


